rename: seperate types.ts file
