+++ src/components/seller/common/SellerModal.tsx
@@ -1,5 +1,5 @@
 import ModalPortal from '@/components/common/ModalPortal';
-import { SellerModalProps } from '@/types/types';
+import { SellerModalProps } from '@/types/common/types';
 import { useEffect } from 'react';
 import { ModalButton } from './Button';
