+++ src/pages/seller/sellerMyPage/SellerMyProfile.tsx
@@ -11,7 +11,7 @@ import {
   SnackBar,
 } from '@/components';
 import { useLocation, useNavigate } from 'react-router-dom';
-import { NoticeType } from '@/types/types';
+import { NoticeType } from '@/types/common/NoticeType.types';
 import PlusIcon from '@/assets/icon/common/PlusIcon.svg?react';
 import cn from '@/utils/cn';
 
@@ -60,12 +60,14 @@ const SellerMyProfile = ({ children }: { children: ReactNode }) => {
       date: '2025.05.01',
       content:
         '부스터 프로 이번 반응이 너무 좋아서 이틀 연장하기로 했어요 ㅎㅎ 많은 관심 감사합니다!부스터 프로 이번 반응이 너무 좋아서 이틀 연장하기로 했어요 ㅎㅎ 많은 관심 감사합니다!',
+      isPrimary: true,
     },
     {
       id: 1,
       title: '제작 오픈 이벤트',
       date: '2025.05.01',
       content: '부스터 프로 이번 반응이 너무 좋아서 이틀 연장하기로 했어요 ',
+      isPrimary: false,
     },
   ];
