+++ src/pages/user/sellerProfile/SelectionTab.tsx
@@ -4,7 +4,7 @@ import LayoutListButton from '@/assets/icon/common/LayoutListButton.svg?react';
 import LayoutAlbumButton from '@/assets/icon/common/LayoutAlbumButton.svg?react';
 import CheckBoxOff from '@/assets/icon/common/CheckBox16Off.svg?react';
 import CheckBoxOn from '@/assets/icon/common/CheckBox16On.svg?react';
-import { ItemType } from '@/types/types';
+import { ItemType } from '@/types/common/ItemType.types';
 
 const SelectionTab = () => {
   const PRODUCT_LIST: ItemType[] = [
