+++ src/pages/user/sellerProfile/SellerProfile.tsx
@@ -9,7 +9,7 @@ import {
   SellerProfileHeader,
   BottomSheet,
 } from '@/components';
-import { NoticeType } from '@/types/types';
+import { NoticeType } from '@/types/common/NoticeType.types';
 
 const SellerProfile = ({ children }: { children: ReactNode }) => {
   const TABS = [
@@ -38,36 +38,42 @@ const SellerProfile = ({ children }: { children: ReactNode }) => {
       date: '2025.05.01',
       content:
         '부스터 프로 이번 반응이 너무 좋아서 이틀 연장하기로 했어요 ㅎㅎ 많은 관심 감사합니다!부스터 프로 이번 반응이 너무 좋아서 이틀 연장하기로 했어요 ㅎㅎ 많은 관심 감사합니다!',
+      isPrimary: true,
     },
     {
       id: 1,
       title: '제작 오픈 이벤트',
       date: '2025.05.01',
       content: '부스터 프로 이번 반응이 너무 좋아서 이틀 연장하기로 했어요 ',
+      isPrimary: false,
     },
     {
       id: 2,
       title: '입접 이벤트🤔💕',
       date: '2025.05.01',
       content: '부스터 프로 이번 반응이 너무 좋아서 이틀 연장하기로 했어요 ',
+      isPrimary: false,
     },
     {
       id: 3,
       title: '입접 이벤트🤔💕',
       date: '2025.05.01',
       content: '부스터 프로 이번 반응이 너무 좋아서 이틀 연장하기로 했어요 ',
+      isPrimary: false,
     },
     {
       id: 4,
       title: '입접 이벤트🤔💕',
       date: '2025.05.01',
       content: '부스터 프로 이번 반응이 너무 좋아서 이틀 연장하기로 했어요 ',
+      isPrimary: false,
     },
     {
       id: 5,
       title: '입접 이벤트🤔💕',
       date: '2025.05.01',
       content: '부스터 프로 이번 반응이 너무 좋아서 이틀 연장하기로 했어요 ',
+      isPrimary: false,
     },
   ];
