+++ src/types/common/ItemType.types.ts
@@ -1,0 +1,12 @@
+export type ItemType = {
+  itemId: number;
+  title: string;
+  name: string;
+  content?: string;
+  thumbnail: string | null;
+  open: string;
+  deadline: string;
+  scrapped: boolean;
+  extend?: boolean;
+  soldOut?: boolean;
+};
