+++ src/types/common/NoticeType.types.ts
@@ -1,0 +1,7 @@
+export type NoticeType = {
+  id: number;
+  title: string;
+  date: string;
+  content: string;
+  isPrimary: boolean;
+};
