+++ src/types/seller/SellerModalProps.types.ts
@@ -1,0 +1,13 @@
+import { SetStateAction } from 'react';
+
+// 모달 props
+export interface SellerModalProps {
+  text: string;
+  description?: string;
+  leftButtonText?: string;
+  leftButtonClick?: () => void;
+  rightButtonText?: string;
+  rightButtonClick: () => void;
+  onClose?: () => void;
+  setIsModalOpen: React.Dispatch<SetStateAction<boolean>>;
+}
